requirements.txt
# Core ML
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.30.0
datasets>=2.18.0

# Interpretability
pyvene>=0.1.4

# Analysis
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pyarrow>=15.0.0  # for parquet support

# Dev / test
pytest>=8.0.0

# Optional: for Llama Guard classification
# (uncomment when ready)
# llama-guard  # not a real package; handled via transformers pipeline
